+++ R/filterInput.R
@@ -143,11 +143,7 @@ method(filterInput, class_character) <- function(x, ...) {
 ## Method: data.frame ####
 method(filterInput, class_data.frame) <- function(x, ...) {
 	filter_input <- function(x, id, nm) {
-		arg_name_id <- arg_name_input_id(x, ...)
-		arg_name_label <- arg_name_input_label(x, ...)
-		args <- list(x, id, nm)
-		names(args) <- c("x", arg_name_id, arg_name_label)
-		args <- c(args, list(...))
+		args <- c(list(x = x), ._id_label_args(x, id, nm, ...), list(...))
 		do.call("filterInput", args)
 	}
 	do.call(
@@ -254,6 +250,12 @@ call_filter_input <- function(x, .f, ...) {
 }
 
 ._call_filter_input <- function(x, .f, ..., call = caller_env()) {
+	args_prepared <- ._prepare_input_args(x, ..., call = call)
+	._call_input(.f, args_prepared, ...)
+}
+
+# Calls `.f` with `args`, plus any `...` that `.f` accepts and `args` lacks
+._call_input <- function(.f, args, ...) {
 	args_provided <- list(...)
 	function_args <- formalArgs(.f)
 	if (identical(.f, selectizeInput)) {
@@ -262,16 +264,22 @@ call_filter_input <- function(x, .f, ...) {
 			setdiff(formalArgs(selectInput), "selectize")
 		)
 	}
-	args_prepared <- ._prepare_input_args(x, ..., call = call)
 	args <- c(
-		args_prepared,
+		args,
 		args_provided[
 			names(args_provided) %in%
 				function_args &
-				!(names(args_provided) %in% names(args_prepared))
+				!(names(args_provided) %in% names(args))
 		]
 	)
 	do.call(.f, args)
+}
+
+# Named inputId / label arguments for one column of a data.frame
+._id_label_args <- function(x, id, label, ...) {
+	args <- list(id, label)
+	names(args) <- c(arg_name_input_id(x, ...), arg_name_input_label(x, ...))
+	args
 }
 
 # Generic: ._apply_ns ####