+++ tests/testthat/helper-filter_config.R
@@ -1,0 +1,6 @@
+df_config <- data.frame(
+	chr = c("b", "a", "c"),
+	fct = factor(c("hi", "lo", "lo"), levels = c("lo", "hi")),
+	int = c(10L, 9L, 2L),
+	dbl = c(1.5, 2.5, 3.5)
+)